requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.0
Pillow==10.1.0
